Moved tests call to a separate Controller in order to make it more flexible (just in case)
